rename files and Copyright
