+++ portal/code/tractus-x-portal/src/components/semantics/SemanticHub.tsx
