+++ portal/code/tractus-x-portal/src/components/semantics/SemanticModelDetail.tsx
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Microsoft
+// Copyright (c) 2021 T-Systems
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
